[Chore] 🍺 User 패키지에서 Customer 패키지로 변경
