+++ src/main/java/com/schedule_hanaro/server/customer/adapter/CustomerAdapter.java
@@ -1,0 +1,4 @@
+package com.schedule_hanaro.server.customer.adapter;
+
+public class CustomerAdapter {
+}
