+++ src/main/java/com/schedule_hanaro/server/customer/controller/CustomerController.java
@@ -1,0 +1,7 @@
+package com.schedule_hanaro.server.customer.controller;
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class CustomerController {
+}
