+++ src/main/java/com/schedule_hanaro/server/customer/entity/Customer.java
@@ -1,0 +1,20 @@
+package com.schedule_hanaro.server.customer.entity;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Customer {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "customer_id")
+    private Long id;
+
+    private String customerName;
+}
