+++ src/main/java/com/schedule_hanaro/server/customer/exception/CustomerException.java
@@ -1,0 +1,4 @@
+package com.schedule_hanaro.server.customer.exception;
+
+public class CustomerException {
+}
