+++ src/main/java/com/schedule_hanaro/server/customer/message/ErrorCode.java
@@ -1,0 +1,4 @@
+package com.schedule_hanaro.server.customer.message;
+
+public enum ErrorCode {
+}
