+++ src/main/java/com/schedule_hanaro/server/customer/message/SuccessMessage.java
@@ -1,0 +1,4 @@
+package com.schedule_hanaro.server.customer.message;
+
+public enum SuccessMessage {
+}
