+++ src/main/java/com/schedule_hanaro/server/customer/repository/CustomerRepository.java
@@ -1,0 +1,4 @@
+package com.schedule_hanaro.server.customer.repository;
+
+public interface CustomerRepository {
+}
