+++ src/main/java/com/schedule_hanaro/server/customer/service/CustomerService.java
@@ -1,0 +1,7 @@
+package com.schedule_hanaro.server.customer.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class CustomerService {
+}
