--- src/main/java/com/schedule_hanaro/server/user/adapter/UserAdapter.java
@@ -1,4 +1,0 @@
-package com.schedule_hanaro.server.user.adapter;
-
-public class UserAdapter {
-}
