--- src/main/java/com/schedule_hanaro/server/user/controller/UserController.java
@@ -1,7 +1,0 @@
-package com.schedule_hanaro.server.user.controller;
-
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class UserController {
-}
