--- src/main/java/com/schedule_hanaro/server/user/entity/User.java
@@ -1,28 +1,0 @@
-package com.schedule_hanaro.server.user.entity;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "user_id")
-    private Long id;
-
-    private String username;
-
-    private void setId(Long id) {
-        this.id = id;
-    }
-
-    private void setUsername(String username) {
-        this.username = username;
-    }
-}
