--- src/main/java/com/schedule_hanaro/server/user/exception/UserException.java
@@ -1,4 +1,0 @@
-package com.schedule_hanaro.server.user.exception;
-
-public class UserException {
-}
