--- src/main/java/com/schedule_hanaro/server/user/message/SuccessMessage.java
@@ -1,4 +1,0 @@
-package com.schedule_hanaro.server.user.message;
-
-public enum SuccessMessage {
-}
