--- src/main/java/com/schedule_hanaro/server/user/repository/UserRepository.java
@@ -1,4 +1,0 @@
-package com.schedule_hanaro.server.user.repository;
-
-public interface UserRepository {
-}
