--- src/main/java/com/schedule_hanaro/server/user/service/UserService.java
@@ -1,7 +1,0 @@
-package com.schedule_hanaro.server.user.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserService {
-}
